Move student related controller and routes into same namespace
