+++ routes/student/profile/profile.routes.js
@@ -1,13 +1,13 @@
 module.exports = app => {
-  const profileController = require('../../controllers/profile/profile.controller')
+  const profileController = require('../../../controllers/student/profile/profile.controller')
   const profileRouter = require('express').Router({ mergeParams: true });
 
   // Ensure user is logged in for all requests handled by this router
-  const auth = require('../../middleware/auth');
+  const auth = require('../../../middleware/auth');
 
   // Get profile data
   profileRouter.get('/', auth, profileController.profileData)
 
   // Tell express to route all requests directed to /api/projects to the router defined in this file
-  app.use('/api/profile', profileRouter);
+  app.use('/api/student/profile', profileRouter);
 }
