+++ routes/student/projects/projects.routes.js
@@ -1,9 +1,9 @@
 module.exports = app => {
-  const projectsController = require('../../controllers/projects/projects.controller')
+  const projectsController = require('../../../controllers/student/projects/projects.controller')
   const projectRouter = require('express').Router({ mergeParams: true });
 
   // Ensure user is logged in for all requests handled by this router
-  const auth = require('../../middleware/auth');
+  const auth = require('../../../middleware/auth');
 
   // Get all projects for the logged in user
   projectRouter.get('/', auth, projectsController.findAll)
@@ -12,5 +12,5 @@ module.exports = app => {
   projectRouter.get('/:id', auth, projectsController.findOne);
 
   // Tell express to route all requests directed to /api/projects to the router defined in this file
-  app.use('/api/projects', projectRouter);
+  app.use('/api/student/projects', projectRouter);
 }
